+++ include/zollstock/concepts/dimensionless.hpp
@@ -1,8 +1,8 @@
-#ifndef __ZOLLSTOCK_PLANAR_ANGLE_CONCEPT_HPP__
-#define __ZOLLSTOCK_PLANAR_ANGLE_CONCEPT_HPP__
+#ifndef __ZOLLSTOCK_CONCEPTS_DIMENSIONLESS_HPP__
+#define __ZOLLSTOCK_CONCEPTS_DIMENSIONLESS_HPP__
 
 
-#include <zollstock/quantity_concept.hpp>
+#include <zollstock/concepts/quantity.hpp>
 #include <zollstock/unit_concept.hpp>
 
 
@@ -15,16 +15,16 @@ namespace zollstock::inline quantities
         // Kapselung des Aufrufes unit_dimensions in einer Vorlagenvariable wegen eines
         // internen Fehlers des MSVC-Compilers.
         template <typename Quantity>
-        inline constexpr bool planar_angle_v =
+        inline constexpr bool dimensionless_v =
             quantity_c<Quantity> &&
             unit_dimensions(Quantity::unit()) == dimensions::_1;
 
     }
 
     template <typename Quantity>
-    concept planar_angle_c = detail::planar_angle_v<Quantity>;
+    concept dimensionless_v = detail::dimensionless_v<Quantity>;
 
 }
 
 
-#endif //__ZOLLSTOCK_PLANAR_ANGLE_CONCEPT_HPP__
+#endif //__ZOLLSTOCK_CONCEPTS_DIMENSIONLESS_HPP__