+++ include/zollstock/concepts/volume.hpp
@@ -1,10 +1,9 @@
-#ifndef __ZOLLSTOCK_VOLUME_CONCEPT_HPP__
-#define __ZOLLSTOCK_VOLUME_CONCEPT_HPP__
+#ifndef __ZOLLSTOCK_CONCEPTS_VOLUME_HPP__
+#define __ZOLLSTOCK_CONCEPTS_VOLUME_HPP__
 
 
-#include <zollstock/quantity_concept.hpp>
-#include <zollstock/unit_concept.hpp>
-
+#include <zollstock/concepts/quantity.hpp>
+#include <zollstock/concepts/unit.hpp>
 
 namespace zollstock::inline quantities
 {
@@ -27,4 +26,4 @@ namespace zollstock::inline quantities
 }
 
 
-#endif //__ZOLLSTOCK_VOLUME_CONCEPT_HPP__
+#endif //__ZOLLSTOCK_CONCEPTS_VOLUME_HPP__