+++ include/zollstock/unit_one.hpp
@@ -11,14 +11,6 @@ namespace zollstock::units::inline literals
 
     ZOLLSTOCK_DEFINE_LITERALS(1, _1)
 
-
-    template <typename Quantity>
-    concept dimensionless_c = requires
-    {
-        requires quantity_c<Quantity>;
-        requires unit_dimensions(Quantity::unit()) == dimensions::_1;
-    };
-
 }