+++ tests/sources/dimensionless_unit_constant_tests.cpp
@@ -1,4 +1,4 @@
-#include <zollstock/unit_concept.hpp>
+#include <zollstock/concepts/unit.hpp>
 #include <zollstock/unit_one_constant.hpp>
 
 #include <catch2/catch_all.hpp>