+++ tests/sources/header_tests/amount_of_substance_concept_header_test.cpp
@@ -1,4 +1,4 @@
-#include <zollstock/amount_of_substance_concept.hpp>
+#include <zollstock/concepts/amount_of_substance.hpp>
 #include <zollstock/tests/header_info.hpp>
 
 #include <catch2/catch_all.hpp>