+++ tests/sources/header_tests/area_concept_header_test.cpp
@@ -1,4 +1,4 @@
-#include <zollstock/area_concept.hpp>
+#include <zollstock/concepts/area.hpp>
 #include <zollstock/tests/header_info.hpp>
 
 #include <catch2/catch_all.hpp>