+++ tests/sources/header_tests/mass_concept_header_test.cpp
@@ -1,4 +1,4 @@
-#include <zollstock/mass_concept.hpp>
+#include <zollstock/concepts/mass.hpp>
 #include <zollstock/tests/header_info.hpp>
 
 #include <catch2/catch_all.hpp>