+++ tests/sources/header_tests/quantity_concepts_header_test.cpp
@@ -1,4 +1,4 @@
-#include <zollstock/quantity_concepts.hpp>
+#include <zollstock/concepts/quantities.hpp>
 #include <zollstock/tests/header_info.hpp>
 
 #include <catch2/catch_all.hpp>